fs: change log to Error

+++ src/fs/create.js
@@ -7,12 +7,23 @@ const file = path.join(path.dirname(url),'files','fresh.txt')
 
 const create = async () => {
     try {
-        const fd = await fs.open(file);
-        console.log('Error: FS operation failed')
+        // const fd = await fs.open(file);
+        // throw new Error('FS operation failed')
+        const srcexist = await checkPathExists(file)
+        if (srcexist) {
+            throw new Error('FS operation failed')
+        }
+        else {
+            await fs.writeFile(file, 'I am fresh and young')
+        }
     } catch (e) {
-        await fs.writeFile(file, 'I am fresh and young')
-        console.log('Created: fresh.txt')
+        console.log(e)
     }
 };
 
 await create();
+async function checkPathExists(path) {
+    return fs.stat(path)
+        .then((res) => res.isFile())
+        .catch((err) => err.code === 'ENOENT' ? false : err);
+}
